// frontend/src/app/api/model/cardDto.ts
/**
 * MemoQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CardDto { 
    id?: number;
    front?: string;
    back?: string;
    status?: CardDto.StatusEnum;
    box?: number;
    createdAt?: string;
}
export namespace CardDto {
    export const StatusEnum = {
        Inactive: 'INACTIVE',
        Active: 'ACTIVE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}