// frontend/src/app/api/model/updateCardRequest.ts
/**
 * MemoQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateCardRequest { 
    front?: string;
    back?: string;
    status?: UpdateCardRequest.StatusEnum;
    box?: number;
}
export namespace UpdateCardRequest {
    export const StatusEnum = {
        Inactive: 'INACTIVE',
        Active: 'ACTIVE',
        Archived: 'ARCHIVED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}